+++ src/SFA.DAS.Approvals.UITests/Project/Helpers/API/LearnerDataOuterApiHelper.cs
@@ -1,5 +1,5 @@
 ﻿using RestSharp;
-using SFA.DAS.FrameworkHelpers;
+using SFA.DAS.Approvals.UITests.Project.Helpers.DataHelpers.ApprenticeshipModel;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -24,7 +24,66 @@ namespace SFA.DAS.Approvals.UITests.Project.Helpers.API
             learnerDataOuterApiClient = context.Get<LearnerDataOuterApiClient>();
         }
 
-        internal async Task<string?> FindLearnerKeyByUlnAsync(string resource, string targetUln)
+        public async Task PushNewLearnersDataToAS(List<LearnerDataAPIDataModel> learnersData, int academicYear)
+        {
+            var resource = $"/provider/{learnersData.First().ukprn}/academicyears/{academicYear}/learners";
+            var payload = JsonHelper.Serialize(learnersData).ToString();
+            await learnerDataOuterApiClient.PostNewLearners(resource, payload);
+        }
+
+
+        public async Task<List<LearnerDataAPIDataModel>> ConvertToLearnerDataAPIDataModel(List<Apprenticeship> listOfApprenticeships)
+        {
+            List<LearnerDataAPIDataModel> listOfLearnerData = new List<LearnerDataAPIDataModel>();
+            await Task.Delay(100);
+
+            foreach (var apprenticeship in listOfApprenticeships)
+            {
+                listOfLearnerData.Add(await ConvertToLearnerDataAPIDataModel(apprenticeship));
+            }
+
+            return listOfLearnerData;
+        }
+
+        
+        public async Task CheckApprenticeIsAvailableInApprovedLearnersList(Apprenticeship apprenticeship)
+        {
+            var resource = $"/Learners/providers/{apprenticeship.UKPRN}/academicyears/{apprenticeship.TrainingDetails.AcademicYear}/learners";
+            var learnerKey = await FindLearnerKeyByUlnAsync(resource, apprenticeship.ApprenticeDetails.ULN);
+            var expectedLearningIdKey = apprenticeship.ApprenticeDetails.LearningIdKey.Trim();
+            Assert.AreEqual(learnerKey.Trim(), expectedLearningIdKey, $"LearningIdKey key extracted from db [{expectedLearningIdKey}] differs from api response: [{learnerKey.Trim()}]");
+
+        }
+
+
+        private async Task<LearnerDataAPIDataModel> ConvertToLearnerDataAPIDataModel(Apprenticeship apprenticeship)
+        {
+            LearnerDataAPIDataModel learnerData = new LearnerDataAPIDataModel();
+
+            learnerData.ukprn = apprenticeship.UKPRN.ToString();
+            learnerData.uln = apprenticeship.ApprenticeDetails.ULN;
+            learnerData.firstname = apprenticeship.ApprenticeDetails.FirstName;
+            learnerData.lastname = apprenticeship.ApprenticeDetails.LastName;
+            learnerData.learnerEmail = apprenticeship.ApprenticeDetails.Email;
+            learnerData.dateOfBirth = apprenticeship.ApprenticeDetails.DateOfBirth.ToString("yyyy-MM-dd");
+            learnerData.startDate = apprenticeship.TrainingDetails.StartDate.ToString("yyyy-MM-dd");
+            learnerData.plannedEndDate = apprenticeship.TrainingDetails.EndDate.ToString("yyyy-MM-dd");
+            learnerData.percentageLearningToBeDelivered = apprenticeship.TrainingDetails.PercentageLearningToBeDelivered;
+            learnerData.epaoPrice = apprenticeship.TrainingDetails.EpaoPrice;
+            learnerData.trainingPrice = apprenticeship.TrainingDetails.TrainingPrice;
+            learnerData.agreementId = apprenticeship.EmployerDetails.AgreementId;
+            learnerData.isFlexiJob = apprenticeship.TrainingDetails.IsFlexiJob;
+            learnerData.plannedOTJTrainingHours = apprenticeship.TrainingDetails.PlannedOTJTrainingHours;
+            learnerData.standardCode = apprenticeship.TrainingDetails.StandardCode;
+            learnerData.consumerReference = apprenticeship.TrainingDetails.ConsumerReference;
+
+            await Task.Delay(100);
+            return learnerData;
+
+        }
+
+
+        private async Task<string?> FindLearnerKeyByUlnAsync(string resource, string targetUln)
         {
             int page = 1;
             const int pageSize = 100;
@@ -60,10 +119,29 @@ namespace SFA.DAS.Approvals.UITests.Project.Helpers.API
         }
 
 
+    }
+
+    public class LearnerDataAPIDataModel
+    {
+        public string uln { get; set; }
+        public string ukprn { get; set; }
+        public string firstname { get; set; }
+        public string lastname { get; set; }
+        public string learnerEmail { get; set; }
+        public string dateOfBirth { get; set; }
+        public string startDate { get; set; }
+        public string plannedEndDate { get; set; }
+        public int percentageLearningToBeDelivered { get; set; }
+        public int epaoPrice { get; set; }
+        public int trainingPrice { get; set; }
+        public string agreementId { get; set; }
+        public bool isFlexiJob { get; set; }
+        public int plannedOTJTrainingHours { get; set; }
+        public int standardCode { get; set; }
+        public string consumerReference { get; set; }
 
 
     }
-
 
     internal class LearnerResponse
     {
