+++ src/SFA.DAS.Approvals.UITests/Project/Steps/LearnerDataOuterApiSteps.cs
@@ -13,20 +13,20 @@ using System.Threading.Tasks;
 namespace SFA.DAS.Approvals.UITests.Project.Steps
 {
     [Binding]
-    public class SldIlrSubmissionSteps
+    public class LearnerDataOuterApiSteps
     {
         private readonly ScenarioContext context;
         private readonly CommonStepsHelper commonStepsHelper;
-        private readonly SLDDataPushHelpers sldDataPushHelpers;
+        private readonly LearnerDataOuterApiHelper learnerDataOuterApiHelper;
         private ProviderStepsHelper providerStepsHelper;
         private ApprenticeDataHelper apprenticeDataHelper;
 
-        public SldIlrSubmissionSteps(ScenarioContext _context)
+        public LearnerDataOuterApiSteps(ScenarioContext _context)
         {
             context = _context;
             commonStepsHelper = new CommonStepsHelper(context);
             providerStepsHelper = new ProviderStepsHelper(context);
-            sldDataPushHelpers = new(context);
+            learnerDataOuterApiHelper = new LearnerDataOuterApiHelper(context);
             apprenticeDataHelper = new ApprenticeDataHelper(context);
         }
 
@@ -63,9 +63,9 @@ namespace SFA.DAS.Approvals.UITests.Project.Steps
 
             var academicYear = listOfApprenticeship.FirstOrDefault().TrainingDetails.AcademicYear;
 
-            var listOflearnerData = await sldDataPushHelpers.ConvertToLearnerDataAPIDataModel(listOfApprenticeship);
+            var listOflearnerData = await learnerDataOuterApiHelper.ConvertToLearnerDataAPIDataModel(listOfApprenticeship);
 
-            await sldDataPushHelpers.PushDataToAS(listOflearnerData, academicYear);
+            await learnerDataOuterApiHelper.PushNewLearnersDataToAS(listOflearnerData, academicYear);
         }
 
         [Given(@"Provider sends an apprentice request \(cohort\) to an employer")]
@@ -122,7 +122,7 @@ namespace SFA.DAS.Approvals.UITests.Project.Steps
         {
             var listOfApprenticeship = context.GetValue<List<Apprenticeship>>();
             var academicYear = listOfApprenticeship.FirstOrDefault().TrainingDetails.AcademicYear;
-            await sldDataPushHelpers.CheckApprenticeIsAvailableInApprovedLearnersList(listOfApprenticeship.FirstOrDefault());
+            await learnerDataOuterApiHelper.CheckApprenticeIsAvailableInApprovedLearnersList(listOfApprenticeship.FirstOrDefault());
             
         }
